Compiler daemons use un-serialized execution specs
